+++ Sources/SwiftUIKit/_Deprecated/Loading/DotLoading+Deprecated.swift
@@ -1,0 +1,5 @@
+@available(*, deprecated, renamed: "DotLoadingAnimation")
+public typealias DotAnimation = DotLoadingAnimation
+
+@available(*, deprecated, renamed: "DotLoadingAnimationText")
+public typealias DotAnimationText = DotLoadingAnimationText